A good start
Basic movement is implemented. Gave the character a circle collider to be able to slide off of frustrating corners easier. You'll also see in the code a *100 was added to both the x and y velocity, that was just so that we could have a smaller number for movementSpeed in the editor.