Tweaked Camera and Added Toggle for Options
Tweked the max camera zoom in and out based on some feedback and updated the script so it's just variables at the top that can be tweaked instead of a few numbers. Added a toggle for the R to restart and Esd to exit to get it out of the way while fighting.